Add possibility to set the SNI host for the connections
Sometimes the URI  and the exact host which we want to connect
have different domains. And we need to have a way to set the domain for
the TLS connection which will be used to verify the TLS cert.
The domain from the URI will still be used to create a underlying TCP
connection.

For example:

Target host: example.com
Target host's TLS: example.net

with setting the SNI host to `example.net` we can still easily connect to
`example.com` and get the correct resources as for `example.net`